requirements.txt
flask
tweepy
python-dotenv
pytest
flask-testing
requests
